+++ src/cli/commands/config.rs
@@ -48,6 +48,13 @@ pub enum ConfigArgs {
     /// Manages the repositories in the config
     #[clap(subcommand)]
     Repositories(RepositoriesArgs),
+
+    /// Resets the `Config.toml` to its default configuration
+    Reset {
+        /// Skips a confirmation prompt
+        #[arg(short, long)]
+        yes: bool,
+    },
 }
 
 /// Manages the repositories in the config.
@@ -159,6 +166,7 @@ impl HandleCommand for ConfigArgs {
             ConfigArgs::Repositories(RepositoriesArgs::DisablePrebuilds { id, value, remove_urls }) => {
                 self.handle_disable_prebuilds(config, id, *value, *remove_urls)
             },
+            ConfigArgs::Reset { yes } => self.handle_reset(config, *yes),
         }
     }
 }
@@ -510,5 +518,24 @@ impl ConfigArgs {
         }
 
         true
+    }
+
+    /// Resets the `Config.toml` to its default configuration.
+    fn handle_reset(&self, config: EditableConfig, yes: bool) {
+        println!("The following configuration will be removed:");
+        self.handle_show(config);
+        println!();
+
+        if !yes {
+            let question = "Resetting the configuration cannot be undone, are you sure you want to continue";
+            if ask_user(question, QuestionResponse::No).unwrap_or_exit(1).is_no_or_invalid() {
+                println!("Cancelled configuration reset");
+                return;
+            }
+        }
+
+        let new_config = EditableConfig::default();
+        new_config.save_to(&Config::get_default_path()).unwrap_or_exit_msg("Cannot save new config", 1);
+        println!("{}", "Successfully reset configuration".to_string().bold().green());
     }
 }